requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
Django==3.0.7
django-cors-headers==3.3.0
django-filter==2.3.0
djangorestframework==3.11.0
docker-compose==1.27.4
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.6.1
Markdown==3.2.2
psycopg2==2.8.4
PyJWT==1.7.1
python-dotenv==0.14.0
python-http-client==3.2.7
pytz==2020.1
requests==2.24.0
selenium==3.141.0
sendgrid==6.4.4
soupsieve==2.0.1
sqlparse==0.3.1
starkbank-ecdsa==1.0.0
urllib3==1.25.10
zipp==3.1.0
